Create Finding most common elements
Finding most common elements
Another powerful usage of Counter is finding the most common elements in a list. This can be done with the .most_common() method.

Practice using this now to find the most common species in a penguins list.

Instructions

Import the Counter object from collections.
Create a Counter of the penguins list called penguins_species_counts; use a list comprehension to return the Species of each penguin to the Counter.
Print the three most common species counts.